requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
python-docx>=0.8.11
